add game state duplication, very janky
basically, node.duplicate() in godot is badly designed, but there is no better alternative than implementing duplicating. In our case, node.duplicate() doesn't duplicate resources because it never even checks type, but we can abuse knowledge of how engine works to know it is looping through all variables and just copying them over to new instance, and add custom getsetter to values we want deep copies for. This is hella jank though since if we are not careful we may accidentally break references if we somehow reference during duplication.